Redesign first steps
