+++ .meteor/versions
@@ -51,7 +51,6 @@ htmljs@1.0.10
 http@1.1.8
 ian:accounts-ui-bootstrap-3@1.2.89
 id-map@1.0.8
-infinitedg:winston@0.7.3
 iron:controller@1.0.12
 iron:core@1.0.11
 iron:dynamic-template@1.0.12
@@ -89,6 +88,7 @@ momentjs:moment@2.14.4
 mongo@1.1.12
 mongo-id@1.0.5
 mongo-livedata@1.0.12
+msavin:sjobs@2.0.1
 npm-bcrypt@0.9.1
 npm-mongo@1.5.48
 observe-sequence@1.0.12
