+++ package.json
@@ -37,14 +37,7 @@
       "logger": true,
       "Papa": true,
       "sAlert": true,
-      "Winston": true,
-      "LiveData": true,
-      "AggrData": true,
-      "Sites": true,
-      "LiveSites": true,
       "DataSeries": true,
-      "Exports": true,
-      "AggrEdits": true,
       "DataExporter": true,
       "SimpleSchema": true,
       "channelHash": true,
