+++ server/main.js
@@ -1,0 +1,3 @@
+import '../imports/startup/startup';
+import '../imports/livefeed/livefeed';
+import '../imports/cronjobs/cronJob_push';
